Final prototype refinements from Figma
Refined the prototype for any last changes made in Figma so they match.

I’ve also hidden/commented out the Non-levy apprentice 2022 to 2023 table in the history view until we get confirmation of whether to show PDFs of not. If not then the history will only show 2023/2024 and 2024/2025 digital statements. If PDFs are in, then this table will be unhidden and designed to include the third column “Format”.